Removed Devop package, moved items to mint
