+++ src/main/java/com/everneth/emi/commands/mint/DevopCommand.java
@@ -1,4 +1,4 @@
-package com.everneth.emi.commands.devop;
+package com.everneth.emi.commands.mint;
 
 import co.aikar.commands.BaseCommand;
 import co.aikar.commands.annotation.*;
